# api-service/src/spec/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"HttpStatusCode": {
				"enum": [
					400,
					500,
					404
				],
				"type": "number"
			},
			"CustomErrorType": {
				"enum": [
					0,
					1
				],
				"type": "number"
			},
			"Record_string.unknown_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"ApplicationError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"statusCode": {
						"$ref": "#/components/schemas/HttpStatusCode"
					},
					"type": {
						"$ref": "#/components/schemas/CustomErrorType"
					},
					"details": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Record_string.unknown_"
							},
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							},
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"type": "number",
									"format": "double"
								},
								"type": "array"
							}
						]
					}
				},
				"required": [
					"name",
					"message",
					"type"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"ActionType.TELEGRAM_NOTIFICATION": {
				"enum": [
					"TELEGRAM_NOTIFICATION"
				],
				"type": "string"
			},
			"TelegramNotification": {
				"properties": {
					"type": {
						"$ref": "#/components/schemas/ActionType.TELEGRAM_NOTIFICATION"
					},
					"payload": {}
				},
				"required": [
					"type",
					"payload"
				],
				"type": "object",
				"additionalProperties": true
			},
			"StrategyAction": {
				"$ref": "#/components/schemas/TelegramNotification"
			},
...
			"SignalsDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "Unique ID of the signal",
						"example": 123
					},
					"strategyId": {
						"type": "number",
						"format": "double",
						"description": "Unique ID of the assosiated strategy",
						"example": 123
					},
					"taskId": {
						"type": "number",
						"format": "double",
						"description": "Unique ID of the assosiated task",
						"example": 123
					},
					"asset": {
						"type": "string",
						"description": "Asset for which the signal was produced"
					},
					"marketId": {
						"type": "number",
						"format": "double",
						"description": "Unique ID of the assosiated market",
						"example": 123
					},
					"producedAt": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						],
						"nullable": true,
						"description": "Signal producing date in ISO format"
					},
					"actionsTriggered": {
						"items": {
							"$ref": "#/components/schemas/StrategyAction"
						},
						"type": "array",
						"nullable": true,
						"description": "Actions triggered by signal"
					},
					"metadata": {
						"description": "Any metadata"
					},
					"createdAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						],
						"description": "Signal creation date in ISO format"
					},
					"updatedAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						],
						"description": "Last signal update date in ISO format"
					}
				},
				"required": [
					"id",
					"strategyId",
					"taskId",
					"asset",
					"marketId",
					"producedAt",
					"actionsTriggered",
					"metadata",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Pick_Signal.Exclude_keyofSignal.id-or-createdAt-or-updatedAt__": {
				"properties": {
					"strategyId": {
						"type": "number",
						"format": "double"
					},
					"marketId": {
						"type": "number",
						"format": "double"
					},
					"taskId": {
						"type": "number",
						"format": "double"
					},
					"asset": {
						"type": "string"
					},
					"producedAt": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						]
					},
					"actionsTriggered": {
						"items": {
							"$ref": "#/components/schemas/TelegramNotification"
						},
						"type": "array"
					},
					"metadata": {}
				},
				"required": [
					"strategyId",
					"marketId",
					"taskId",
					"asset"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_Signal.id-or-createdAt-or-updatedAt_": {
				"$ref": "#/components/schemas/Pick_Signal.Exclude_keyofSignal.id-or-createdAt-or-updatedAt__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
...
			"MarketResponseDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "Unique ID of the market",
						"example": 123
					},
					"name": {
						"type": "string",
						"description": "Market name"
					},
					"url": {
						"type": "string",
						"description": "Market url"
					},
					"createdAt": {
						"type": "string",
						"description": "Market creation date in ISO format"
					},
					"updatedAt": {
						"type": "string",
						"description": "Last market update date in ISO format"
					}
				},
				"required": [
					"id",
					"name",
					"url",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Pick_Market.name-or-url_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"url"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "api-service",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/signals": {
			"post": {
				"operationId": "CreateSignal",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SignalsDto"
								}
							}
						}
					},
					"400": {
						"description": "Invalid market id",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApplicationError"
								}
							}
						}
					}
				},
				"description": "Create a new signal",
				"summary": "Returns created signal",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_Signal.id-or-createdAt-or-updatedAt_"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "SearchSignals",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SignalsDto"
									},
									"type": "array"
								}
							}
						}
					},
					"400": {
						"description": "Invalid payload",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApplicationError"
								}
							}
						}
					}
				},
				"description": "Get array of signals",
				"summary": "Returns signals array filtered by query params",
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "ids",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "number",
								"format": "double"
							}
						}
					},
					{
						"in": "query",
						"name": "strategyIds",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "number",
								"format": "double"
							}
						}
					},
					{
						"in": "query",
						"name": "taskIds",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "number",
								"format": "double"
							}
						}
					},
					{
						"in": "query",
						"name": "assets",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "marketIds",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "number",
								"format": "double"
							}
						}
					},
					{
						"in": "query",
						"name": "from",
						"required": false,
						"schema": {}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "to",
						"required": false,
						"schema": {}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"ASC",
								"DESC"
							]
						}
					}
				]
			}
		},
		"/signals/{id}": {
			"get": {
				"operationId": "GetSignalById",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SignalsDto"
								}
							}
						}
					},
					"404": {
						"description": "Signal is not found.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApplicationError"
								}
							}
						}
					}
				},
				"description": "Get a signal by id",
				"summary": "Returns signal by id",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteSignalById",
				"responses": {
					"204": {
						"description": ""
					},
					"404": {
						"description": "Signal is not found.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApplicationError"
								}
							}
						}
					}
				},
				"description": "Delete signal by id",
				"summary": "Deletes signal by it's id, returns empty response",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/markets": {
			"post": {
				"operationId": "CreateMarket",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MarketResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Invalid payload",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApplicationError"
								}
							}
						}
					}
				},
				"description": "Create a new market",
				"summary": "Returns created market",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Pick_Market.name-or-url_"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "SearchMarkets",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MarketResponseDto"
									},
									"type": "array"
								}
							}
						}
					},
					"400": {
						"description": "Invalid payload",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApplicationError"
								}
							}
						}
					}
				},
				"description": "Get array of markets",
				"summary": "Returns markets array filtered by query params",
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "ids",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "number",
								"format": "double"
							}
						}
					},
					{
						"in": "query",
						"name": "names",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "url",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "from",
						"required": false,
						"schema": {}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "to",
						"required": false,
						"schema": {}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"ASC",
								"DESC"
							]
						}
					}
				]
			}
		},
		"/markets/{id}": {
			"get": {
				"operationId": "GetMarketById",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MarketResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "Market is not found.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApplicationError"
								}
							}
						}
					}
				},
				"description": "Get a market by id",
				"summary": "Returns market by id",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteMarketById",
				"responses": {
					"204": {
						"description": ""
					},
					"404": {
						"description": "Market is not found.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApplicationError"
								}
							}
						}
					}
				},
				"description": "Delete market by id",
				"summary": "Deletes market by it's id, returns empty response",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateMarketById",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MarketResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "Market is not found.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApplicationError"
								}
							}
						}
					}
				},
				"description": "Update market by id",
				"summary": "Returns updated market",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Pick_Market.name-or-url_"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}